refactor(oracledb): move runtime symbols to consumers
